requirements.txt
requests
PyYAML
python-dotenv
bleak